pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore api"
authors = [
    {name = "danielcruz",email = "danielcscruz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
